WIP: check for changed restore file
